Configure CI on Ascent

See merge request exasgd/frameworks/exago!90

+++ build.sh
@@ -16,17 +16,27 @@ cleanup() {
 
 trap 'cleanup $? $LINENO' EXIT
 
-# set -xv
+set -xv
 make_args="-j 8"
 ctest_args=" -VV "
 export CXXFLAGS='-I/share/apps/magma/2.5.2/cuda10.2/include -I/share/apps/cuda/10.2/include/'
 export OMPI_MCA_btl="^vader,tcp,openib,uct"
+export BUILD=1
+export TEST=1
 
 while [[ $# -gt 0 ]]; do
   case $1 in
-  -s|--short)
+  --short)
     echo "Only running representative subset of unit tests."
     ctest_args="$ctest_args -R UNIT_TEST"
+    shift
+    ;;
+  --build-only)
+    export TEST=0 BUILD=1
+    shift
+    ;;
+  --test-only)
+    export TEST=1 BUILD=0
     shift
     ;;
   *)
@@ -49,26 +59,83 @@ module purge
 
 case $MY_CLUSTER in
 newell*)
+  source /etc/profile.d/modules.sh
+  module purge
   export OMP_CANCELLATION=true
   export OMP_PROC_BIND=true
   export OMPI_MCA_pml="ucx"
   export UCX_NET_DEVICES=mlx5_1:1,mlx5_3:1
   export MY_CLUSTER=newell
+  export PROJ_DIR=/qfs/projects/exasgd
+  export APPS_DIR=/share/apps
   module load gcc/7.4.0
   module load cmake/3.16.4
   module load openmpi/3.1.5
   module load magma/2.5.2_cuda10.2
   module load metis/5.1.0
   module load cuda/10.2
+  export MY_PETSC_DIR=$PROJ_DIR/$MY_CLUSTER/petsc
+  export MY_UMPIRE_DIR=$PROJ_DIR/$MY_CLUSTER/umpire
+  export MY_RAJA_DIR=$PROJ_DIR/$MY_CLUSTER/raja
+  export MY_HIOP_DIR=$PROJ_DIR/$MY_CLUSTER/hiop
+  export MY_UMPIRE_DIR=$PROJ_DIR/$MY_CLUSTER/umpire
+  export MY_IPOPT_DIR=$PROJ_DIR/$MY_CLUSTER/ipopt
+  export MY_MAGMA_DIR=$APPS_DIR/magma/2.5.2/cuda10.2
+  export NVBLAS_CONFIG_FILE=$PROJ_DIR/$MY_CLUSTER/nvblas.conf
   ;;
 dl|shared|marianas)
+  source /etc/profile.d/modules.sh
+  module purge
   export MY_CLUSTER=marianas
+  export PROJ_DIR=/qfs/projects/exasgd
+  export APPS_DIR=/share/apps
   module load gcc/7.3.0
   module load cmake/3.15.3
   module load openmpi/3.1.3
   module load magma/2.5.2_cuda10.2
-  module load metis/5.1.0
   module load cuda/10.2.89
+  module load metis/5.1.0
+  export MY_PETSC_DIR=$PROJ_DIR/$MY_CLUSTER/petsc
+  export MY_UMPIRE_DIR=$PROJ_DIR/$MY_CLUSTER/umpire
+  export MY_RAJA_DIR=$PROJ_DIR/$MY_CLUSTER/raja
+  export MY_HIOP_DIR=$PROJ_DIR/$MY_CLUSTER/hiop
+  export MY_UMPIRE_DIR=$PROJ_DIR/$MY_CLUSTER/umpire
+  export MY_IPOPT_DIR=$PROJ_DIR/$MY_CLUSTER/ipopt
+  export MY_MAGMA_DIR=$APPS_DIR/magma/2.5.2/cuda10.2
+  export NVBLAS_CONFIG_FILE=$PROJ_DIR/$MY_CLUSTER/nvblas.conf
+  ;;
+ascent)
+  export MY_CLUSTER=ascent
+  export PROJ_DIR=/gpfs/wolf/proj-shared/csc359
+  module purge
+  module load cuda/11.0.2
+  module use $PROJ_DIR/$MY_CLUSTER/Modulefiles/Core
+  module load exasgd-base
+  module load gcc-ext/7.4.0
+  module load spectrum-mpi-ext
+  module load magma/2.5.3-cuda11
+  module load raja
+  module load umpire
+  module load openblas
+  module load cmake/3.18.2
+  export MY_RAJA_DIR=$RAJA_ROOT
+  export MY_UMPIRE_DIR=$UMPIRE_ROOT
+  export MY_MAGMA_DIR=$MAGMA_ROOT
+  export MY_PETSC_DIR=$PROJ_DIR/$MY_CLUSTER/petsc-3.13.5
+  export MY_HIOP_DIR=$PROJ_DIR/installs/hiop
+  export MY_NVCC_ARCH="sm_70"
+  extra_cmake_args="$extra_cmake_args -DHIOP_NVCC_ARCH=$MY_NVCC_ARCH"
+  extra_cmake_args="$extra_cmake_args -DEXAGO_TEST_WITH_BSUB=ON"
+  if [[ ! -f $builddir/nvblas.conf ]]; then
+    cat > $builddir/nvblas.conf <<-EOD
+	NVBLAS_LOGFILE  nvblas.log
+	NVBLAS_CPU_BLAS_LIB  /gpfs/wolf/proj-shared/csc359/ascent/Compiler/gcc-7.4.0/openblas/0.3.10/lib/libopenblas.so
+	NVBLAS_GPU_LIST ALL
+	NVBLAS_TILE_DIM 2048
+	NVBLAS_AUTOPIN_MEM_ENABLED
+	EOD
+  fi
+  export NVBLAS_CONFIG_FILE=$builddir/nvblas.conf
   ;;
 *)
   echo
@@ -80,38 +147,28 @@ esac
 
 builddir=$(pwd)/build
 installdir=$(pwd)/install
-petsc_dir=/qfs/projects/exasgd/$MY_CLUSTER/petsc
-hiop_dir=/qfs/projects/exasgd/$MY_CLUSTER/hiop
-raja_dir=/qfs/projects/exasgd/$MY_CLUSTER/raja
-umpire_dir=/qfs/projects/exasgd/$MY_CLUSTER/umpire
-magma_dir=/share/apps/magma/2.5.2/cuda10.2/
-metis_dir=/share/apps/metis/5.1.0/
-cuda_dir=/share/apps/cuda/10.2/
-export NVBLAS_CONFIG_FILE=/qfs/projects/exasgd/$MY_CLUSTER/nvblas.conf
 export LD_LIBRARY_PATH="$magma_dir/lib:$hiop_dir/lib:$LD_LIBRARY_PATH"
 export CTEST_OUTPUT_ON_FAILURE=1
 
 module list
 
-declare -a builds=(
-  " \
+cmake_args=" \
   -DCMAKE_INSTALL_PREFIX=$installdir/ \
   -DCMAKE_BUILD_TYPE=Debug \
   -DEXAGO_ENABLE_GPU=ON \
   -DEXAGO_ENABLE_HIOP=ON \
-  -DEXAGO_ENABLE_IPOPT=ON \
+  -DEXAGO_ENABLE_IPOPT=OFF \
   -DEXAGO_ENABLE_MPI=ON \
   -DEXAGO_ENABLE_PETSC=ON \
   -DEXAGO_RUN_TESTS=ON \
   -DEXAGO_ENABLE_RAJA=ON \
-  -DRAJA_DIR=$raja_dir \
-  -Dumpire_DIR=$umpire_dir \
-  -DHIOP_DIR=$hiop_dir \
-  -DIPOPT_DIR=/qfs/projects/exasgd/$MY_CLUSTER/ipopt \
-  -DMAGMA_DIR=$magma_dir \
-  -DPETSC_DIR=$petsc_dir"
-)
-
+  -DRAJA_DIR=$MY_RAJA_DIR \
+  -Dumpire_DIR=$MY_UMPIRE_DIR \
+  -DHIOP_DIR=$MY_HIOP_DIR \
+  -DIPOPT_DIR=$MY_IPOPT_DIR \
+  -DMAGMA_DIR=$MY_MAGMA_DIR \
+  -DPETSC_DIR=$MY_PETSC_DIR \
+  $extra_cmake_args"
 
 #  NOTE: The following is required when running from Gitlab CI via slurm job
 base_path=`dirname $0`
@@ -119,13 +176,8 @@ if [ -z "$SLURM_SUBMIT_DIR" ]; then
     cd $base_path          || exit 1
 fi
 
-for ((i=0; i<${#builds[@]}; i++))
-do
-  echo
-  echo Build $((i+1)) / ${#builds[@]}
-  echo
-  args=${builds[i]}
-  echo Building with args $args
+if [[ $BUILD -eq 1 ]]; then
+  echo Building with args $cmake_args
 
   [ -d $builddir ] && rm -rf $builddir
   mkdir -p $builddir
@@ -148,7 +200,7 @@ do
   echo
   echo Configuring
   echo
-  cmake $args .. || exit 1
+  cmake $cmake_args .. || exit 1
 
   echo
   echo Building
@@ -159,16 +211,16 @@ do
   echo Installing
   echo
   make install || exit 1
+  popd
+fi
 
+if [[ $TEST -eq 1 ]]; then
+  pushd $builddir
   echo
   echo Testing
   echo
   ctest $ctest_args || exit 1
   popd
-
-  echo
-  echo Build $((i+1)) / ${#builds[@]} successful
-  echo
-done
+fi
 
 exit 0
